refactor: inputHandler/outputHandler 코드 분리
