+++ src/domain/Race.js
@@ -5,16 +5,19 @@ import {
   displayResultTitle,
   displayWinner,
 } from "../io/OutputHandler.js";
+import { validateCarNames } from "../utils/validate.js";
+import Car from "./Car.js";
 
 export const start = async () => {
-  const cars = await getCarNames();
+  const cars = await getCars();
   const attempt = await getAttempt();
   displayResultTitle();
   for (let i = 0; i < attempt; i++) {
     moveCars(cars);
     displayRaceResult(cars);
   }
-  displayWinner(cars);
+  const winners = determineWinners(cars);
+  displayWinner(winners);
 };
 
 const moveCars = (cars) => {
@@ -25,4 +28,26 @@ const moveCars = (cars) => {
 
 const getRandomNumber = () => {
   return Math.floor(Math.random() * (MAX_RANDOM_VALUE - MIN_RANDOM_VALUE + 1));
+};
+
+const getCars = async () => {
+  while (true) {
+    const name = await getCarNames();
+    const carNames = name.split(",");
+    try {
+      let cars = carNames.map((name) => new Car(name));
+      validateCarNames(cars, carNames);
+      return cars;
+    } catch (error) {
+      console.log(error.message);
+    }
+  }
+};
+
+const determineWinners = (cars) => {
+  const maxPosition = Math.max(...cars.map((car) => car.position));
+  const winners = cars
+    .filter((car) => car.position === maxPosition)
+    .map((car) => car.name);
+  return winners;
 };
